MinBias version of basic plots code
You know all those fancy hMinBias and hHydjet histos I made? I didn't
bother to fill them, instead using the same exact histogram names as I
did in the jet triggered version of code, so this code has a LOT of
extra lines of histos that are *never filled or written or plotted*.
But it works at least.